clip_service/requirements.txt
flask
torch
transformers
Pillow
requests
gunicorn
